requirements.txt
adjustText==0.8
bcrypt==3.2.2
cached-property==1.5.2
cffi==1.15.1
cryptography==3.4.6
cycler==0.11.0
cytoolz==0.12.0
fonttools==4.37.1
future==0.18.2
gensim==4.2.0
install==1.3.5
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
natsort==8.2.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
paramiko==2.10.1
Pillow==9.2.0
plotly==5.11.0
psutil==5.8.0
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
smart-open==6.2.0
tenacity==8.1.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.0.dev20220913
typing_extensions==4.3.0
watchdog==2.0.1
